Apply feedback, move tests
